DATAGRAPH-567 - NullPointerException when saving relationship entity with end node field in camel-case
