+++ neo4j-ogm/src/test/java/org/neo4j/ogm/integration/RelationshipEntityTest.java
@@ -18,8 +18,12 @@
 
 package org.neo4j.ogm.integration;
 
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.neo4j.helpers.collection.IteratorUtil;
+import org.neo4j.ogm.domain.cineasts.annotated.Actor;
+import org.neo4j.ogm.domain.cineasts.annotated.Knows;
 import org.neo4j.ogm.domain.friendships.Friendship;
 import org.neo4j.ogm.domain.friendships.Person;
 import org.neo4j.ogm.exception.InvalidDepthException;
@@ -29,17 +33,18 @@ import org.neo4j.ogm.session.SessionFactory;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Date;
 
 import static org.junit.Assert.*;
 
-public class RelationshipEntityHydrationTest extends InMemoryServerTest {
+public class RelationshipEntityTest extends InMemoryServerTest {
 
     private static SessionFactory sessionFactory;
 
     @Before
     public void init() throws IOException {
         setUp();
-        sessionFactory = new SessionFactory("org.neo4j.ogm.domain.friendships");
+        sessionFactory = new SessionFactory("org.neo4j.ogm.domain.friendships", "org.neo4j.ogm.domain.cineasts.annotated");
         session = sessionFactory.openSession("http://localhost:" + neoPort);
     }
 
@@ -297,5 +302,28 @@ public class RelationshipEntityHydrationTest extends InMemoryServerTest {
         session.load(Friendship.class, friendship.getId(),0);
     }
 
+
+    /**
+     * @see DATAGRAPH-567
+     */
+    @Test
+    public void shouldSaveRelationshipEntityWithCamelCaseStartEndNodes() {
+        Actor bruce = new Actor("Bruce");
+        Actor jim = new Actor("Jim");
+
+        Knows knows = new Knows();
+        knows.setFirstActor(bruce);
+        knows.setSecondActor(jim);
+        knows.setSince(new Date());
+
+        bruce.getKnows().add(knows);
+
+        session.save(bruce);
+
+        Actor actor = IteratorUtil.firstOrNull(session.loadByProperty(Actor.class, new Property<String, Object>("name", "Bruce")));
+        Assert.assertNotNull(actor);
+        assertEquals(1,actor.getKnows().size());
+        assertEquals("Jim",actor.getKnows().iterator().next().getSecondActor().getName());
+    }
 
 }
